+++ packages/ui/src/components/NotificationItem/variants.ts
@@ -1,0 +1,11 @@
+export type Variant =
+  | "preExercise"
+  | "postExercise"
+  | "exerciseConfirm"
+  | "reserve"
+  | "cancel"
+  | "edit"
+  | "connect"
+  | "disconnect"
+  | "deny"
+  | "session";
